feat: remove jruby support
BREAKING CHANGE: remove jruby support
