+++ .gitignore
@@ -7,9 +7,6 @@
 /spec/reports/
 /tmp/
 /.ruby-version
-/.mvn
 
 # rspec failure tracking
 .rspec_status
-
-*.jar
