+++ Rakefile
@@ -5,9 +5,4 @@ require "rspec/core/rake_task"
 
 RSpec::Core::RakeTask.new(:spec)
 
-if RUBY_PLATFORM == "java"
-  require "warbler"
-  Warbler::Task.new
-end
-
 task default: :spec
