+++ ckb-sdk-ruby.gemspec
@@ -2,8 +2,6 @@ lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 require "ckb/version"
 
-is_jruby = RUBY_PLATFORM == 'java'
-
 Gem::Specification.new do |spec|
   spec.name          = "ckb-sdk-ruby"
   spec.version       = CKB::VERSION
@@ -44,13 +42,6 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "pry", "~> 0.12.2"
 
   spec.add_dependency "net-http-persistent", "~> 3.0.0"
-
-  if !is_jruby
-    spec.add_dependency "rbnacl", "~> 6.0", ">= 6.0.1"
-    spec.add_dependency "bitcoin-secp256k1", "~> 0.5.0"
-  else
-    spec.add_development_dependency "warbler", "~> 2.0"
-
-    spec.requirements << "jar org.bouncycastle, bcprov-jdk15on, 1.61"
-  end
+  spec.add_dependency "rbnacl", "~> 6.0", ">= 6.0.1"
+  spec.add_dependency "bitcoin-secp256k1", "~> 0.5.0"
 end
