+++ lib/ckb.rb
@@ -13,8 +13,6 @@ require "ckb/wallet"
 require "ckb/address"
 require "ckb/key"
 
-require "ckb-sdk-ruby_jars" if RUBY_PLATFORM == "java"
-
 module CKB
   class Error < StandardError; end
   # Your code goes here...
